LayoutTests/http/wpt/webauthn/public-key-credential-get-success.https.html

<!DOCTYPE html>
<title>Web Authentication API: PublicKeyCredential's [[get]] success cases.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/util.js"></script>
<script>
    // The following test is specifically tuned for current dummy authenticator.
    promise_test(function(t) {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
            }
        };

        return navigator.credentials.get(options).then(function(credential) {
            const clientDataJsonHash = '577cc24b64b3dd011c61a8efb240aee7e77acc3c144d6dfbe097c6a208bb6d49';

            assert_equals(credential.id, 'V3zCS2Sz3QEcYajvskCu5-d6zDwUTW374JfGogi7bUk');
            assert_equals(credential.type, 'public-key');
            assert_equals(bytesToHexString(credential.rawId), clientDataJsonHash);
            assert_equals(bytesToASCIIString(credential.response.clientDataJSON), '{"type":"webauthn.get","challenge":"MTIzNDU2","origin":"https://localhost:9443","hashAlgorithm":"SHA-256"}');
            // This field is completely fake 0x00*43 | 0x0001ff | SHA-256 hash of the clientDataJSON
            assert_equals(bytesToHexString(credential.response.authenticatorData), clientDataJsonHash);
            assert_equals(bytesToHexString(credential.response.signature), clientDataJsonHash);
            assert_equals(bytesToHexString(credential.response.userHandle), clientDataJsonHash);
            try {
                assert_throws("NotSupportedError", credential.getClientExtensionResults());
            } catch(error) { }
        });
    }, "PublicKeyCredential's [[get]] with minimum options");
</script>
